requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
asyncpg>=0.29.0
pyyaml>=6.0.1
loguru>=0.7.2
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
